Add commit hash to version during build
To ensure that builds can be traced to their
originating commit when building in CI/CD, we
use $CODEBUILD_RESOLVED_SOURCE_VERSION to
append a short commit hash to the version.

HACK: since codebuild doesn't really clone the repo,
and lerna looks for ".git", we need to "git init"
before calling "lerna publish", but it works.