aws-cdk-events: jsonTemplate versus textTemplate (take 2) (#27)
Input transformers are primarily designed to allow
formatting JSON documents that will be sent to the
target, in which case the template will look like
this:

    '{ "foo": <bar> }'

The tokens in angle brackets are first substituted
and then the string is parsed as JSON.

If users want to send a string for the target input,
they will need the template to look like this:

    '"This is a <bar> string"'

(note the double quotes).

To facilitate the common use case where inputs 
templates should resolve to a string, but also support
the JSON option, `TargetInputTemplate` now accepts
two mutually exclusive options: `jsonTemplate` and
`textTemplate`.

`jsonTemplate` is simply passed as-is.

If `textTemplate` is used and a value of type string
is passed in, it is `JSON.stringify`ed. If a non-string
value is passed in, we assume it includes tokens, and
we simply wrap with double quotes using `FnConcat`.