Change Custom Resource mechanism to make Lambda explicit
Introducing `SingletonLambda` which can be instantiated multiple times
but only adds the Lambda to the underlying stack once.

A CustomResource now explicitly takes either a `LambdaRef` or a `TopicRef`,
and the Lambda can be a SingletonLambda.

This splitting up makes it possible to establish bidirectional permissions
between the Custom Resource Lambda and other resources.