Fix `cdk init` for private beta phase
Installation of NPM dependencies was broken due to the packages not
being published in a repository. This changes introduces a `cdk-npm`
command that uses `verdaccio` to provide a local NPM repository that
contains the required packages for CDK to work, and proxies other
packages to the central NPM repository.

This also changes the `cdk init` invocation to use the `--language`
option (which can optionally be set via `~/.cdk.json`), and accepts
a template name as a positional argument (defaulting to the `app`
template). A `--list` flag was added to allow listing the templates.

Fixes #99