Upgrade jsii (support statics) (#90)
Update jsii distribution to jsii-0.4.0-pre+6248947 following
added support for static methods and properties (awslabs/jsii#26).

Minor fixes required to adhere to jsii support for
static methods and properties:

Rename RESOURCE_TYPE_NAME and RESOURCE_PROPERTIES
of the property registry to camelCase.

Return `any` in Resource.attribute decoration
since jsii doesn't support lambda functions.

Add tslint.json to .gitigonre since now it is created
by jsii (if it doesn't already exist).

To maintain traceability with the official jsii build
used, we now check in the bundle zip file instead of
the tarballs and expand them in ./install.sh to `./.local-npm` 
(which is not checked-in)