Also support AWS_DEFAULT_REGION environment variable
We use the account from the environment variables, but only look
for the region in ~/.aws/config directory, which is inconsistent.
This is the official environment variable to use for regions.

Also, for some reason on my machine, the config setting from the
environment variable is not even used, which makes running integ
tests fail. This fixes that.