Remove --pinentry-mode option from gpg invokation
It is not supported by the version of `gnupg` (`1.4.16`) that is built
into the latest superchain image (prior image used `gnupg >= 2.1`).

The `--pinentry-mode` option is used to allow the `gpg-agent` process
to request the passphrase from the process that invoked it, which is
required when using `gnupg >= 2.1` and passing a private key's
passphrase using the `--passphrase-fd` option (since the fd is owned
by the invoker process).