Introducing `y-npm` (aka: Why NPM?) (#162)
A drop-in replacement for the `npm` command that uses a local repository
as a source for packages, and falls back to the standard `npm` registry
for packages that are not locally available.

This can be used to make private packages available to consumers without
the need to set up a full blown repository and managing user
authentication and authorization; and also to overlay customized
versions of third party packages in a simple way.

The `y-npm` command has an additional sub-command `y-npm y-ls`
that will output the name and version of all packages that are locally
available.

This commit also replaces cdk-beta-npm in favor of using y-npm. Also,
the release .zip file no longer includes a full installation of the cdk,
instead only makes y-npm, so people can use y-npm instead of npm and
work as they normally would besides that (this reduces the file size of
the release zip by about 50%).