Make execution of `pack.sh` much faster
Since the `pack.sh` script is supposed to be run after `build.sh` was
executed, the `prepare` script of all packages was already performed,
and  it is possible to only run the `prepublish` script, then execute
`npm pack --ignore-scripts` in parallel.

This removes the double-compilation of all packages (including
slow-to-build packages such as `aws-cdk-docs` or `@aws-cdk/resources`,
that involve expensive code generation).