Store default account in disk cache
Given an access key ID, the account ID is fixed.
This means that we don't need to look up the default
account every time the toolkit is used.

This change will store the mapping between access key
and account ID in `~/.cdk/cache/accounts.json`
and will fetch the account ID from there if it exists.