Insert AWS::CDK::Metadata resource (#221)
Collect the list of loaded NodeJS modules and determine their names
and versions from locating their `package.json` file. If the `private`
attribute is set, ignore the module.

Added a `runtime` entry to the CX API for synthesizing stacks,
so as to be able to pass information from the application back
to the toolkit.

Added a documentation section about version reporting, to inform the
users of what we are collecting, why, and how they can opt out of it.