aws-cdk-java: clean up pom.xml and include in deployed maven repo
Include the pom.xml file in the maven repo, so that transitive dependencies will be respected.

This also required us to clean up the file and not include the path to the jsii-runtime repo. Instead, we simply deploy it into ~/.m2 during `./generate.sh`.

Not ideal, but works for now. When jsii is published to Maven Central, this is all going to go away, so trying to not over-engineer it.