Java init template (#229)
The template includes a java shim for the toolkit called "app.sh". The shim depends on a file `.classpath.txt` that's generated by maven during build. It points to `target/classes` for classpath. This supports a workflow where you can iterate in the IDE (and run unit tests), but then go to the command line to execute the toolkit, and you will target the same classes the IDE compiled.

BETA: the template includes a pom.xml file that depends on `~/.cdk/repo/maven` as a pragmatic solution for beta. There's value in adding the repository in the pom.xml file directly as this gets us seamless IDE experience that works out of the box.

This change also modifies the pom.xml file for aws-cdk-java to not explicitly include the local jsii repository, which is temporary anyway until we publish to Maven Central. This allows us to include the full pom.xml in the maven package, so transitive dependencies work.