Introduce @aws-cdk/cloudformation-resource-spec
The package provides a "convenient" way to access the CloudFormation and
SAM resource specification, with all CDK patches applied to it, and
surfaces this as a rich object model. Convenience methods are provided
to make it easier to access sub-sets of the specification; for example
in order to generate code for a particular L1/L2 namespace.

The package includes a series of tests that validate the produced
specification actually matches the expected object model.