Java Getting Started (+ rewrite without "cdk init")
Add getting started and welcome in Java (and JavaScript).

Rewrite "Getting Started" to detail the process of defining CDK apps
without using `cdk init`. This is in order to improve the understanding
of users when they onboard the CDK, which fixes #219.

Fix TOC to include only two layers deep which fixes #129.