Handle return from yargs command modules
Prevent the application from reaching the code path in `main` when the
command was handled directly by a `commandDir` module. This prevents
the operation from succeeding, while displaying a message such as
*"Unknown command: docs"* (#256).

In order to achieve this, the module decorator sets `argv.result` on
successful execution of the command module's handler; and if
`argv.result` is set, `main` does not get called.

Fixes #256