[BREAKING] Move `@aws-cdk/resources` classes into L2 packages (#264)
The `create_missing_libraries.sh` script was used to bootstrap L2
packages for all the namespaces that didn't have one already.

Existing L2 packages were manually edited to:
- Invoke `cfn2ts` to generate the L1 class library for the service/scope
- Depend on `@aws-cdk/runtime` (which used to be blended in
  `@aws-cdk/resources`, and was extracted out)
- Export the L1 library as expected
- Generated resource and property classes are namespaced under 'cloudformation'

The `@aws-cdk/assert` behavior was changed to rely on
`@aws-cdk/cdk-cfnspec` (formerly `@aws-cdk/cloudformation-resource-spec`)
to determine the `UpdateType` of attributes, rendering the `registry`
mechanism that was introduced to `@aws-cdk/resources` a while
ago redundant (hence it got dropped).

This required temporarily disabling use of '--reject-cycles' on 'lerna bootstrap'
as it would blow up on the cycle created (see #293).