Centralize build logic, enforce code coverage (#337)
BUILD

Reorganize build. This change introduces a new 'cdk-build-tools' package used
by all packages. All commands for all packages are now 'cdk-build', 'cdk-test',
etc. This centralizes the copy/pasted logic and provides a single point where
we can provide better Windows compatibility.

The 'cdk-build' script detects when sources are unchanged since the last
successful build and skips running the TypeScript compiler in that case,
which saves time if you only changed a couple of packages.

The 'cdk-integ' scripts are moved from @aws-cdk/assert to a new package
called cdk-integ-tools, otherwise the packages that the integ tools
depend could not be built using cdk-build-tools.

The build tools include a fingerpriting system where they'll calculate a
hash of a source tree and skip rebuilding if the hashes haven't changed.
An opportunistic rebuild is automatically triggered as part of
'cdk-test'.

COVERAGE

Code coverage is now checked for all packages using 'nyc'. The global config 
lives in the cdk-build-tools package.

Some packages have unit tests but don't have enough coverage. For those close
enough, add a couple tests to bump the coverage over the threshold.  For those
too far away, the thresholds are lowered per package. Those should be fixed as
soon as possible.

SECRETS

Add allowed git-secrets pattern for AWS account '123456789012', which is the
one account number allowed for use in tests.

ADDITIONAL FIXES

`pkglint` now properly checks the contents of the `test/` directory in
order to enforce the presence of the NPM testing commands.

Fixes #270