Remove dependency from toolkit on CDK (#352)
Due to a cyclic dependency caused by the need for CDK libraries to depend on the toolkit for testing purposes, we decided not to use the CDK to define the "toolkit stack". It's a very simple stack that essentially contains a single bucket.

Therefore, we just embed the template in the toolkit code, and finally we don't have cycles anymore.

Tested this manually by:
* Bootstrapping the toolkit stack in my dev account
* Updating the toolkit stack
* Verifying that "deploy" can discover the bucket and upload a template

The empty "WaitHandle" stack also used the CDK, so this was also embedded inline.

--reject-cycles is enabled now

Fixes #293