Assets
Assets represent local files or directories which can be bundled as part
of CDK constructs. During deployment, the toolkit will upload assets
to the "Toolkit Bucket", and use CloudFormation Parameters to reference
the asset in deploy-time.

Assets expose the following deploy-time attributes:

 * s3BucketName
 * s3ObjectKey
 * s3Url

Furthermore, the `asset.grantRead(principal)` will add IAM read permissions
for the asset to the desired principal.