Attach 'error' metadata and fail synthesis when context cannot be resolved
Instead of throwing an error when context cannot be resolved, we return
dummy values but attach a metadata entry "aws:cdk:error" to the construct
with a message explaining that context cannot be resolved.

The toolkit picks up these metadata entries and will fail synthesis,
unless --ignore-errors is set.

The benefit is that this enables unit test use cases, where users are okay
with dummy contextual information.

INFO messages can also be emitted now using `context.addInfo(message)`.

The metadata keys for info, warning and errors are normalized to use
the "aws:cdk" prefix, and formally defined in cxapi.