Context providers: return dummy values if env is undefined, but emit errors to fail "cdk synth" (#227)
Throwing if "env" is undefined makes it hard to unit test stacks that
use context providers. It requires dummy values for account and region,
which is boilerplate and non-intuitive.

Since we already have a concept of dummy context values, this change
short-circuits context resolution and returns the dummy values in case
the stack's env is undefined.

To ensure dummy values are not used in production, when a dummy value
is returned, we attach a metadata entry `aws:cdk:error` to the construct
with a message explaining that context cannot be resolved.

The toolkit picks up these metadata entries and will fail synthesis,
unless --ignore-errors is set.

INFO messages can also be emitted now using `context.addInfo(message)`.

The metadata keys for info, warning and errors are normalized to use
the "aws:cdk" prefix, and formally defined in cxapi.