Combine Connections and DefaultConnections classes
The DefaultConnections type could be advertised on a class because
changing the type of the 'connections: IConnections' property was not
allowed in C#. Hence, all uses of this member needed to be cast.

Change the interface to always advertise the concrete class
'Connections', and fold the methods of DefaultConnections in there.
They now have a runtime check on the presence of default port and
will throw if the Connections class hasn't been initialized properly.

This change is breaking for middleware writers, not for direct
consumers.

This fixes #418.