jsii-pacmak packaging (#468)
Setup a "package" npm script in all modules (via pkglint) to perform
packaging for the module:

If this is a jsii module, jsii-pacmak is invoked, which now generates
and builds the module for all target languages. Output goes to `dist/npm`,
`dist/java` and `dist/dotnet`.

If this is not a jsii module, we invoke 'npm pack' and move the tarball
to `dist/npm` (so it looks like the jsii output).

aws-cdk-docs dist/ directory modified to only contain a `docs/` output
(staging moved out).

Private modules don't have a "package" script.

Updated pack.sh to simply call "lerna run package", which will result
in "dist/" directories with artifacts throughput the repo.

Now, we merge all dist/ directories into "./pack/" and bundle-beta.sh
will use it as the base for the distribution zip file (implemented
in `pack-collect.sh`)

The distribution zip file still supports y-npm based installation, but
it's structure is now:


    aws-cdk-${version}.zip
    │
    ├─ bin
    ├─ node_modules
    ├─ y
    │  └─ npm - y-npm repository for local installs (duplicate tarballs)
    │
    ├─ docs   - rendered docsite
    ├─ npm    - npm tarballs
    ├─ dotnet - nuget packages
    ├─ java   - maven repository
    │
    └─ .version


Fixes #111

Also:
* Don't build java example for now, we need to turn it into an integration test.
* Bundle the jsii java runtime w/ other java packages