Change pkglint rules with respects to repository URL & author (#497)
Adds an `organization: true` property to the author so we can appropriately label the
author entry in `pom.xml` files (this is an extension defined/used by `jsii`).

Switch the `repository.url` to the `http` github endpoint, as this allows us to produce a
"standard" URL for use in `pom.xml` files for Maven Central. While this isn't "required"
per se, it makes the package metadata "better" in that they adhere to the Maven Central
best practices (repository URL is a web interface) while keeping the repository URL
something that can be used for read-only source access.