feat(cdk): Tagging support via TagManager (#538)
Fixes #91, Closes #458 (as obsolete)

The TagManager is a class Construct authors can use to implement tagging
consistently. The manager provides the ability to propagate tags from
parents to child, override parent tags in the child, set default tags on
the child that can be overwritten by parents, and block tags from
parents.

Adding tagging support for the Vpc and Subnet constructs.