Add support for CloudFormation actions in CodePipeline
---
Author's note: This work mostly belongs to Max Hall (@hallmaxw)
It was at some point commented out during a refactor - this is just adding it back.
Most of the credit should go to him.
---

Anywho - this adds the ability for customers to use CloudForamtion actions in CodePipeline.
This follows the same pattern as previous CodePipeline actions by making it it's own package.