fix(aws-lambda): java - invalid cast for inline LambdaRuntime members (#505)
TypeScript uses inference to determine if a cast to an interface is legal but
strongly-typed languages like Java require that the down-casted class will
explicitly implement the interface.

The JavaRuntime class has static members that are down-casted from JavaRuntime
to a set of interfaces, to allow strong-typing of properties for various Lambda
use cases. These down-casts fail in e.g. Java because JavaRuntime doesn't implement
these interfaces explicitly.

We should add a compile-time check in jsii for such a use case.

Fixes #504