Detect and correct "incorrect" use of `new FnJoin`
A common developer error is to pass an array as the second argument of
`new FnJoin`, but that argument is variadic. This causes invalid
CloudFormation templates to be generated. This change detects this usage
pattern (vararg contains a single element that is an array) and fixes
it up transparently.

Related to #512