feat(cdk): construct.uniqueId and relax construct id constraints
Adds a property `uniqueId` to Construct which returns an 
alphanumeric 255-length-limited tree-unique identity for a construct.

Relax constraints for construct id (previously known as "name") to only
restrict the usage of the path separator. Otherwise, all characters are
allowed. This will allow using the construct id for a wider range of
purposes, but since logical IDs (and uniqueId now) are alpha-safe, it's 
okay.

Deprecate `construct.name` in favor of `construct.id`.