feat(aws-sns): bucket notification destination (#560)
Allow SNS topics to be used as bucket notification destinations.

To avoid taking a dependency on aws-s3, extracted the bucket notification destination API
into a separate module @aws-cdk/aws-s3-notifications, which only includes the required
interfaces.

We still take a devDependency on s3, but that's fine.

Added examples/** to the global .nycrc