feat(aws-sqs): support bucket notifications (#564)
Allow sqs.Queue to be used as a destination for bucket notifications.
Also supports queues that are server-side encrypted using a custom KMS key.

Added support for specifying KMS keys in `QueueRef`, including export/import support.

Added support to `kms.EncryptionKey` to allow forbidding "no-op" when
adding policy statements. This is needed in order to fail the user from
subscribing a queue to receive bucket notifications without the ability
to allow S3 to read the confirmation messages from the queue as described
in the S3 documentation [1]

[1]: https://docs.aws.amazon.com/AmazonS3/latest/dev/ways-to-add-notification-config-to-bucket.html