feat(cdk-build-tools): Allow configuring jsii, jsii-pacmak and tsc
Allows additional options to be used to customize the behavior of `cdk-build-tools` commands, making
it significantly easier to test a new version of `jsii`, `jsii-pacmak` or `tsc` without having to make complex
changes in the codebase or using `npm link`.

Permits using environment variables to specify those options. The environment variables are using the
`TRUMP_CASED` name of the command and option, separated with an `_` character. For example,
one can use `CDK_BUILD_JSII` to configure a different `jsii` executable to the `cdk-build` command.

The default behavior is unchanged from the current (not setting any of those will use the commands
provided by the dependency closure of the package).