fix(aws-lambda): jsii use no long requires 'sourceAccount' (#728)
Change the typing of 'sourceAccount' to be optional string, so that the
typing over JSII will be properly inferred. The current type is 'any' so
that we can pass in Tokens, but since the Token stringification change
typing as a string no longer forms an obstacle, and in fact it will be
more inline with upcoming type system changes.

Fixed #714.