fix(aws-s3): remove `policy` argument
This argument served as a red herring, and made it harder
for people to find the correct method of setting a bucket
policy, by calling `addToPolicy()`.

The provided method of providing an existing Policy object on object
creation cannot actually work, because a `BucketPolicy` object requires
the `Bucket` instance as a constructor argument, which means the
`Bucket` object cannot take a `BucketPolicy` in its constructor.

Fixes #672.