feat: use stringified tokens for resource attributes instead of strong types (#712)
This change removes all strong-types generated to represent 
CloudFormation resource attributes that return a string  (such 
as `QueueArn`, `DeploymentGroupId`) and replaces them 
with `string` that contains a stringified token (via #518).

This allows working with these attributes as if they were regular
strings, and dramatically simplifies the type-system and unneeded
wrapping when assigning values to such attributes.

The ".ref" property of each resource has been replaced with
a property named according to the actual semantic meaning
of the intrinsic reference (such as `queueArn`), and 
also returns a `string`.

Users can test if a string contains stringified tokens using the function 
`unresolved(o)`, which can be applied to either strings or objects 
(previously was `isToken(o)`).

Fixes #695 (opened #744 to follow up on non-string attributes)