Allow use of a RepositoryRef when creating a CodeCommitSoureAction (#758)
When trying to use an imported CodeCommit repository (e.g. from a separate stack), one needs to
be able to create the source action using a `RepositoryRef` and not a `Repository`. In order for this
to work, the repository clone URLs needed to be made available on the `RepositoryRef` class.