Introduce StackAwareToken, make all pseudos Stack-Aware.
Introduce context in resolve(), so that StackAwareToken can do something
else if it's a Token from a different stack.

Introduce a two-phase Freeze protocol which is a definite moment at
which all lazy tokens must be resolved (so that the StackAwareTokens
can build cross-Stack links).

Change lock() into the freezing protocol, remove unlock as this was
a speculative feature that is unused.