feat: Manage IAM permissions for (some) CFN CodePipeline actions (#843)
When adding CloudFormation actions to CodePipeline, the pipeline's role must be
granted appropriate permissions on the CloudFormation stacks in order for the
pipeline to work. This adds the relevant permission management to the ChangeSet
actions (`CreateReplaceChangeSet`, `ExecuteChangeSet`).

A bonus BREAKING CHANGE is that the `Artifact.subartifact` method of the
CodePipeline API was renamed to `Artifact.atPath`.