refactor: remove app boilerplate and improvements to cx protocol (#868)
Change the cx protocol to use environment variables and file output
instead of argv/stdout. This simplifies the App boilerplate code.
Now, apps look like this:

    const app = new App();
    // add stacks
    app.run();

The protocol was also simplified to basically always synthesize all
stacks in the app, eliminating the need for a "cx request" altogether.

The toolkit was modified to cache the response of the first successful
execution of the app (after environment contextual is resolved). This
reduces the number of app executions to a maximum of 2 (if env context
is missing) and normally 1. Previoiusly, "list" and "synth" were two
separate commands, so 2 was the normal case.

The protocol now uses the following environment variables:

 * `CDK_OUTDIR` represents the synthesis output directory. The toolkit
   allocates a temporary working directory when it executes the app and
   sets this variable.
 * `CDK_CONTEXT_JSON` is a JSON stringified context object.

At the moment, if `CDK_OUTDIR` is not defined, an error message is printed
indicating the minimum version of the toolkit required to interact with
the app.

Fixes #216

BREAKING CHANGE

This is a major breaking change:

- The `cdk.App` initializer doesn't accept any arguments.
- The `cdk.App#run` method does not return a `string` anymore.

All AWS CDK apps in all languages would need to be modified
to adhere to the new API of the `cdk.App` construct.

Instead of:

    const app = new App(process.argv); // ERROR
    // add stacks
    process.stdout.write(app.run());   // ERROR

The new usage is:

    const app = new App();
    // add stacks
    app.run();

In order to interact with applications written using this
version, the CDK Toolkit must also be update using:

    $ npm i -g aws-cdk