fix: Emit valid YAML-1.1 (#876)
`yamljs` would emit YAML version 1.2, but CloudFormation expects YAML version 1.1. The differences
between versions include some surrogate representations of booleans (`ON`, `OFF`, `YES`, ...) that are
no longer supported in 1.2. This changes implementation to `js-yaml`, which makes extra effort in
emitting YAML that is compatible with previous versions of the standard.

Fixes #875